Moves FrameCache code into drake/multibody/parsing/sdf
