+++ drake/examples/double_pendulum/BUILD.bazel
@@ -12,16 +12,15 @@ load("//tools/lint:lint.bzl", "add_lint_tests")
 drake_cc_library(
     name = "sdf_helpers",
     srcs = [
-        "frame_cache.cc",
         "sdf_helpers.cc",
     ],
     hdrs = [
-        "frame_cache.h",
         "sdf_helpers.h",
     ],
     deps = [
         "//drake/common:essential",
         "//drake/multibody:rigid_body_tree",
+        "//drake/multibody/parsing/sdf:frame_cache",
         "@eigen",
         "@sdformat",
     ],
@@ -66,13 +65,6 @@ drake_cc_library(
         "//drake/multibody:rigid_body",
         "//drake/multibody/shapes",
         "@gtest//:without_main",
-    ],
-)
-
-drake_cc_googletest(
-    name = "frame_cache_test",
-    deps = [
-        ":sdf_helpers",
     ],
 )
