+++ drake/examples/double_pendulum/sdf_helpers.cc
@@ -10,7 +10,7 @@
 #include "drake/common/drake_assert.h"
 #include "drake/common/drake_throw.h"
 #include "drake/common/eigen_types.h"
-#include "drake/examples/double_pendulum/frame_cache.h"
+#include "drake/multibody/parsing/sdf/frame_cache.h"
 #include "drake/multibody/joints/fixed_joint.h"
 #include "drake/multibody/joints/revolute_joint.h"
 #include "drake/multibody/rigid_body.h"
