+++ drake/multibody/parsing/sdf/frame_cache.h
